Switching to 'concurrent.futures' because it's so much easier than asyncio.
I tried some other Python parallel primitives and came away
disappointed. Python is like PowerShell in that there are too many
parallel implementations, all distinct and not compatible.

threading works (especially for this case), but it's for I/O bound tasks
only. You also can't return results from Thread objects, and the
approaches to do so are kinda messy.

multiprocessing looks nice, but it pickles and unpickles everythings, so
having shared objects for concurrency tracking is difficult.

asyncio, as this change describes, doesn't work nicely on some platforms
and it is way, way too complicated.

concurrent.futures seems to work perfectly, but it's buried beneath the
other approaches.