Implementing genre-based exclusion.
'playlist.py' is getting rather large, but the data structure isn't that
complicated, just the operations on it.

Genre-based exclusion still won't work quite right if it is enabled
while on an excluded song, but that'll be work for another time.

All that's left now is cleanup (including the above) and full
randomization across playlists.